Added crude lighting demonstration

+++ PixelGameEngineTinker/PixelGameEngineTinker/Settings.h
@@ -29,40 +29,39 @@ public:
 	// Screen
 	struct Screen
 	{
-		static const std::uint64_t SCREEN_RESOLUTION_X = 1920;
-		static const std::uint64_t SCREEN_RESOLUTION_Y = 1080;
+		static const std::uint16_t SCREEN_RESOLUTION_X = 1920;
+		static const std::uint16_t SCREEN_RESOLUTION_Y = 1080;
 
-		static const std::uint64_t PIXEL_SIZE = 1;
-		static const std::uint64_t SCREEN_PIXEL_WIDTH = SCREEN_RESOLUTION_X / PIXEL_SIZE;
-		static const std::uint64_t SCREEN_PIXEL_HEIGHT = SCREEN_RESOLUTION_Y / PIXEL_SIZE;
-		static const std::uint64_t CELL_PIXEL_SIZE = 16 / PIXEL_SIZE;
+		static const std::uint16_t PIXEL_SIZE = 1;
+		static const std::uint16_t SCREEN_PIXEL_WIDTH = SCREEN_RESOLUTION_X / PIXEL_SIZE;
+		static const std::uint16_t SCREEN_PIXEL_HEIGHT = SCREEN_RESOLUTION_Y / PIXEL_SIZE;
+		static const std::uint16_t CELL_PIXEL_SIZE = 16 / PIXEL_SIZE;
 
-		static const std::uint64_t SCREEN_CELL_WIDTH = SCREEN_PIXEL_WIDTH / CELL_PIXEL_SIZE;
-		static const std::uint64_t SCREEN_CELL_HEIGHT = SCREEN_PIXEL_HEIGHT / CELL_PIXEL_SIZE;
+		static const std::uint16_t SCREEN_CELL_WIDTH = SCREEN_PIXEL_WIDTH / CELL_PIXEL_SIZE;
+		static const std::uint16_t SCREEN_CELL_HEIGHT = SCREEN_PIXEL_HEIGHT / CELL_PIXEL_SIZE;
 	};
 
 
 	// World
 	struct World
 	{
-		static const std::uint64_t CHUNK_CELL_SIZE = 32; // chunk cell size ( n x n ) // [hardcode]
-		static const std::uint64_t NUM_CELLS_PER_CHUNK = CHUNK_CELL_SIZE * CHUNK_CELL_SIZE;
-		static const std::uint64_t CHUNK_RADIUS = 2;
+		static const std::uint16_t CHUNK_CELL_SIZE = 32; // chunk cell size ( n x n ) // [hardcode]
+		static const std::uint16_t NUM_CELLS_PER_CHUNK = CHUNK_CELL_SIZE * CHUNK_CELL_SIZE;
+		static const std::uint16_t CHUNK_RADIUS = 2;
 
-		static const std::uint64_t MAX_SAVED_CHUNKS = 50; // Maximum number of chunks in queue to save before discarding
-		static const std::uint64_t SPRITE_TILE_REFRESH_RATE = 300; // Every 5 minutes, refresh the tileDecals if it has not yet from condition variable notification
+		static const std::uint16_t MAX_SAVED_CHUNKS = 50; // Maximum number of chunks in queue to save before discarding
+		static const std::uint16_t SPRITE_TILE_REFRESH_RATE = 300; // Every 5 minutes, refresh the tileDecals if it has not yet from condition variable notification
 	};
 
 
 	// WorldChunk
 	struct WorldChunk
 	{
-		static const std::uint64_t SIZE = World::CHUNK_CELL_SIZE;
+		static const std::uint16_t SIZE = World::CHUNK_CELL_SIZE;
 
-
-		static const std::uint64_t NUM_TILE_RENDERS = 341; // 4^0 + 4^1 + 4^2 + ... + 4^n ( for n = log( chunkSize ) = quadTreeMaxLevel ) // [hardcode]
-		static const std::uint32_t TILE_RENDER_MIN_LEVEL = 0; // [ hard code ]
-		static const std::uint32_t TILE_RENDER_MAX_LEVEL = 4; // [ hard code ]
+		static const std::uint16_t NUM_TILE_RENDERS = 341; // 4^0 + 4^1 + 4^2 + ... + 4^n ( for n = log( chunkSize ) = quadTreeMaxLevel ) // [hardcode]
+		static const std::uint16_t TILE_RENDER_MIN_LEVEL = 0; // [ hard code ]
+		static const std::uint16_t TILE_RENDER_MAX_LEVEL = 4; // [ hard code ]
 		static const long double TILE_RENDER_MIN_CELL_SIZE; // [ hard code ]
 	};
 
@@ -70,24 +69,24 @@ public:
 	// QuadTree
 	struct QuadTree
 	{
-		static const std::uint64_t MIN_LEVEL = 0;  // [hardcode]
-		static const std::uint64_t MAX_LEVEL = 4; // log2( CHUNK_CELL_SIZE ) - 1  // [hardcode] [locked due to tile png not having 256x256]
+		static const std::uint16_t MIN_LEVEL = 0;  // [hardcode]
+		static const std::uint16_t MAX_LEVEL = 4; // log2( CHUNK_CELL_SIZE ) - 1  // [hardcode] [locked due to tile png not having 256x256]
 	};
 
 
 	// Camera
 	struct Camera
 	{
-		static const std::uint64_t FOCAL_POINT_CELL_WIDTH = World::CHUNK_CELL_SIZE;
-		static const std::uint64_t FOCAL_POINT_CELL_HEIGHT = World::CHUNK_CELL_SIZE;
+		static const std::uint16_t FOCAL_POINT_CELL_WIDTH = World::CHUNK_CELL_SIZE;
+		static const std::uint16_t FOCAL_POINT_CELL_HEIGHT = World::CHUNK_CELL_SIZE;
 
-		static const std::uint64_t VIEW_CELL_WIDTH = Screen::SCREEN_CELL_WIDTH;
-		static const std::uint64_t VIEW_CELL_HEIGHT = Screen::SCREEN_CELL_HEIGHT;
+		static const std::uint16_t VIEW_CELL_WIDTH = Screen::SCREEN_CELL_WIDTH;
+		static const std::uint16_t VIEW_CELL_HEIGHT = Screen::SCREEN_CELL_HEIGHT;
 
-		static const std::int64_t ABSOLUTE_PIXEL_OFFSET_X = Screen::CELL_PIXEL_SIZE * ( Screen::SCREEN_CELL_WIDTH / 2 ) - Screen::CELL_PIXEL_SIZE * ( FOCAL_POINT_CELL_WIDTH / 2 );
-		static const std::int64_t ABSOLUTE_PIXEL_OFFSET_Y = Screen::CELL_PIXEL_SIZE * ( Screen::SCREEN_CELL_HEIGHT / 2 ) - Screen::CELL_PIXEL_SIZE * ( FOCAL_POINT_CELL_HEIGHT / 2 );
+		static const std::int32_t ABSOLUTE_PIXEL_OFFSET_X = Screen::CELL_PIXEL_SIZE * ( Screen::SCREEN_CELL_WIDTH / 2 ) - Screen::CELL_PIXEL_SIZE * ( FOCAL_POINT_CELL_WIDTH / 2 );
+		static const std::int32_t ABSOLUTE_PIXEL_OFFSET_Y = Screen::CELL_PIXEL_SIZE * ( Screen::SCREEN_CELL_HEIGHT / 2 ) - Screen::CELL_PIXEL_SIZE * ( FOCAL_POINT_CELL_HEIGHT / 2 );
 
-		static const std::uint32_t CONSOLIDATED_TILE_OFFSET = 128;
+		static const std::uint16_t CONSOLIDATED_TILE_OFFSET = 128;
 	};
